feat(dashboard): add speaking page and update listening page
- added `speaking.tsx` page under dashboard features
- updated `app/dashboard/speaking/page.tsx` to use new structure
- improved `listening.tsx` page for consistency
