+++ features/dashboard/pages/listening.tsx
@@ -280,7 +280,7 @@ export const ListeningDashboardPage = () => {
               </Button>
             </div>
 
-            <div className="flex flex-col gap-[10px]">
+            <div className="flex flex-col gap-2.5">
               <p className="text-sm text-[#dedede]">
                 Based on {accurationType} accuracy
               </p>
